Fix wrong threads flag extracting for docker commands
