+++ main.go
@@ -1,20 +1,11 @@
 package main
 
 import (
-	"encoding/json"
-	"errors"
 	"fmt"
-	"github.com/jfrog/gofrog/version"
-	"github.com/jfrog/jfrog-client-go/http/httpclient"
-	"github.com/jfrog/jfrog-client-go/utils/errorutils"
-	"github.com/jfrog/jfrog-client-go/utils/io/httputils"
-	"net/http"
 	"os"
-	"path"
 	"runtime"
 	"sort"
 	"strings"
-	"time"
 
 	"github.com/agnivade/levenshtein"
 	corecommon "github.com/jfrog/jfrog-cli-core/v2/docs/common"
@@ -74,11 +65,6 @@ OPTIONS:
 {{end}}
 `
 
-type githubResponse struct {
-	TagName string `json:"tag_name,omitempty"`
-	URL     string `json:"html_url"`
-}
-
 func main() {
 	log.SetDefaultLogger()
 	err := execMain()
@@ -127,7 +113,7 @@ func execMain() error {
 	app.Before = func(ctx *cli.Context) error {
 		clientlog.Debug("JFrog CLI version:", app.Version)
 		clientlog.Debug("OS/Arch:", runtime.GOOS+"/"+runtime.GOARCH)
-		warningMessage, err := checkNewCliVersionAvailable(app.Version)
+		warningMessage, err := cliutils.CheckNewCliVersionAvailable(app.Version)
 		if err != nil {
 			clientlog.Debug("failed while trying to check latest JFrog CLI version:", err.Error())
 		}
@@ -138,68 +124,6 @@ func execMain() error {
 	}
 	err := app.Run(args)
 	return err
-}
-
-// Checks if the requested plugin exists in registry and does not exist locally.
-func checkNewCliVersionAvailable(currentVersion string) (warningMessage string, err error) {
-	shouldCheck, err := shouldCheckLatestCliVersion()
-	if err != nil || !shouldCheck {
-		return
-	}
-	githubVersionInfo, err := getLatestCliVersionFromGithubAPI()
-	if err != nil {
-		return
-	}
-	latestVersion := strings.TrimPrefix(githubVersionInfo.TagName, "v")
-	if version.NewVersion(latestVersion).Compare(currentVersion) < 0 {
-		warningMessage = strings.Join([]string{
-			fmt.Sprintf("You are using JFrog CLI version %s, however version %s is available.", coreutils.PrintComment(currentVersion), coreutils.PrintTitle(latestVersion)),
-			fmt.Sprintf("To install the latest version, visit: %sgetcli", coreutils.JFrogComUrl),
-			"To see the release notes, visit: " + githubVersionInfo.URL,
-			fmt.Sprintf("To ignore this message you can use %s=TRUE", cliutils.JfrogCliAvoidNewVersionWarning),
-		},
-			"\n")
-	}
-	return
-}
-
-func shouldCheckLatestCliVersion() (shouldCheck bool, err error) {
-	if strings.ToLower(os.Getenv(cliutils.JfrogCliAvoidNewVersionWarning)) == "true" {
-		return
-	}
-	homeDir, err := coreutils.GetJfrogHomeDir()
-	if err != nil {
-		return
-	}
-	indicatorFile := path.Join(homeDir, "Latest_Cli_Version_Check_Indicator")
-	fileInfo, err := os.Stat(indicatorFile)
-	if err != nil && !os.IsNotExist(err) {
-		err = fmt.Errorf("couldn't get indicator file %s info: %s", indicatorFile, err.Error())
-		return
-	}
-	if err == nil && (time.Now().UnixMilli()-fileInfo.ModTime().UnixMilli()) < cliutils.LatestCliVersionCheckInterval.Milliseconds() {
-		// Timestamp file exists and updated less than 3 hours ago, therefor no need to check version again
-		return
-	}
-	return true, os.WriteFile(indicatorFile, []byte{}, 0666)
-}
-
-func getLatestCliVersionFromGithubAPI() (githubVersionInfo githubResponse, err error) {
-	client, err := httpclient.ClientBuilder().Build()
-	if err != nil {
-		return
-	}
-	resp, body, _, err := client.SendGet("https://api.github.com/repos/jfrog/jfrog-cli/releases/latest", true, httputils.HttpClientDetails{HttpTimeout: time.Second * 2}, "")
-	if err != nil {
-		err = errors.New("couldn't get latest JFrog CLI latest version info from GitHub API: " + err.Error())
-		return
-	}
-	err = errorutils.CheckResponseStatusWithBody(resp, body, http.StatusOK)
-	if err != nil {
-		return
-	}
-	err = json.Unmarshal(body, &githubVersionInfo)
-	return
 }
 
 // Detects typos and can identify one or more valid commands similar to the error command.
